HAUKI-486 refactoring language utils
